+++ src/main/java/org/opensearch/knn/index/codec/KNN80Codec/KNN80DocValuesConsumer.java
@@ -7,73 +7,36 @@ package org.opensearch.knn.index.codec.KNN80Codec;
 
 import lombok.NonNull;
 import lombok.extern.log4j.Log4j2;
-import org.apache.lucene.store.ChecksumIndexInput;
 import org.opensearch.common.StopWatch;
-import org.opensearch.common.xcontent.XContentHelper;
-import org.opensearch.core.common.bytes.BytesArray;
-import org.opensearch.core.xcontent.MediaTypeRegistry;
-import org.opensearch.core.xcontent.DeprecationHandler;
-import org.opensearch.core.xcontent.NamedXContentRegistry;
-import org.opensearch.knn.index.SpaceType;
-import org.opensearch.knn.index.util.IndexUtil;
-import org.opensearch.knn.index.KNNSettings;
-import org.opensearch.knn.index.VectorDataType;
-import org.opensearch.knn.index.codec.transfer.VectorTransfer;
-import org.opensearch.knn.index.codec.transfer.VectorTransferByte;
-import org.opensearch.knn.index.codec.transfer.VectorTransferFloat;
-import org.opensearch.knn.jni.JNIService;
-import org.opensearch.knn.index.codec.util.KNNCodecUtil;
 import org.opensearch.knn.index.engine.KNNEngine;
-import org.opensearch.knn.indices.Model;
 import org.opensearch.knn.indices.ModelCache;
-import org.opensearch.knn.plugin.stats.KNNCounter;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.lucene.codecs.DocValuesConsumer;
 import org.apache.lucene.codecs.DocValuesProducer;
-import org.apache.lucene.index.BinaryDocValues;
 import org.apache.lucene.index.DocValuesType;
 import org.apache.lucene.index.FieldInfo;
 import org.apache.lucene.index.MergeState;
 import org.apache.lucene.index.SegmentWriteState;
-import org.apache.lucene.store.FSDirectory;
-import org.apache.lucene.store.FilterDirectory;
+import org.opensearch.knn.index.codec.nativeindex.NativeIndexWriter;
 import org.opensearch.knn.index.mapper.KNNVectorFieldMapper;
 import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.plugin.stats.KNNGraphValue;
 
-import java.io.Closeable;
 import java.io.IOException;
-import java.io.OutputStream;
-import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.nio.file.StandardOpenOption;
-import java.security.AccessController;
-import java.security.PrivilegedAction;
-import java.util.HashMap;
-import java.util.Map;
 
-import static org.apache.lucene.codecs.CodecUtil.FOOTER_MAGIC;
 import static org.opensearch.knn.common.KNNConstants.MODEL_ID;
-import static org.opensearch.knn.common.KNNConstants.PARAMETERS;
-import static org.opensearch.knn.index.codec.util.KNNCodecUtil.buildEngineFileName;
-import static org.opensearch.knn.index.codec.util.KNNCodecUtil.calculateArraySize;
-import static org.opensearch.knn.index.engine.faiss.Faiss.FAISS_BINARY_INDEX_DESCRIPTION_PREFIX;
 
 /**
  * This class writes the KNN docvalues to the segments
  */
 @Log4j2
-class KNN80DocValuesConsumer extends DocValuesConsumer implements Closeable {
+class KNN80DocValuesConsumer extends DocValuesConsumer {
 
     private final Logger logger = LogManager.getLogger(KNN80DocValuesConsumer.class);
 
     private final DocValuesConsumer delegatee;
     private final SegmentWriteState state;
-
-    private static final Long CRC32_CHECKSUM_SANITY = 0xFFFFFFFF00000000L;
 
     KNN80DocValuesConsumer(DocValuesConsumer delegatee, SegmentWriteState state) {
         this.delegatee = delegatee;
@@ -113,175 +76,7 @@ class KNN80DocValuesConsumer extends DocValuesConsumer implements Closeable {
 
     public void addKNNBinaryField(FieldInfo field, DocValuesProducer valuesProducer, boolean isMerge, boolean isRefresh)
         throws IOException {
-        // Get values to be indexed
-        BinaryDocValues values = valuesProducer.getBinary(field);
-        final KNNEngine knnEngine = getKNNEngine(field);
-        final String engineFileName = buildEngineFileName(
-            state.segmentInfo.name,
-            knnEngine.getVersion(),
-            field.name,
-            knnEngine.getExtension()
-        );
-        final String indexPath = Paths.get(
-            ((FSDirectory) (FilterDirectory.unwrap(state.directory))).getDirectory().toString(),
-            engineFileName
-        ).toString();
-
-        // Determine if we are creating an index from a model or from scratch
-        NativeIndexCreator indexCreator;
-        KNNCodecUtil.Pair pair;
-        Map<String, String> fieldAttributes = field.attributes();
-        VectorDataType vectorDataType;
-
-        if (fieldAttributes.containsKey(MODEL_ID)) {
-            String modelId = fieldAttributes.get(MODEL_ID);
-            Model model = ModelCache.getInstance().get(modelId);
-            if (model.getModelBlob() == null) {
-                throw new RuntimeException(String.format("There is no trained model with id \"%s\"", modelId));
-            }
-            vectorDataType = model.getModelMetadata().getVectorDataType();
-            pair = KNNCodecUtil.getPair(values, getVectorTransfer(vectorDataType));
-            indexCreator = () -> createKNNIndexFromTemplate(model, pair, knnEngine, indexPath);
-        } else {
-            // get vector data type from field attributes or provide default value
-            vectorDataType = VectorDataType.get(
-                fieldAttributes.getOrDefault(KNNConstants.VECTOR_DATA_TYPE_FIELD, VectorDataType.DEFAULT.getValue())
-            );
-            pair = KNNCodecUtil.getPair(values, getVectorTransfer(vectorDataType));
-            indexCreator = () -> createKNNIndexFromScratch(field, pair, knnEngine, indexPath);
-        }
-
-        // Skip index creation if no vectors or docs in segment
-        if (pair.getVectorAddress() == 0 || pair.docs.length == 0) {
-            logger.info("Skipping engine index creation as there are no vectors or docs in the segment");
-            return;
-        }
-
-        long arraySize = calculateArraySize(pair.docs.length, pair.getDimension(), vectorDataType);
-
-        if (isMerge) {
-            KNNGraphValue.MERGE_CURRENT_OPERATIONS.increment();
-            KNNGraphValue.MERGE_CURRENT_DOCS.incrementBy(pair.docs.length);
-            KNNGraphValue.MERGE_CURRENT_SIZE_IN_BYTES.incrementBy(arraySize);
-            recordMergeStats(pair.docs.length, arraySize);
-        }
-
-        // Increment counter for number of graph index requests
-        KNNCounter.GRAPH_INDEX_REQUESTS.increment();
-
-        if (isRefresh) {
-            recordRefreshStats();
-        }
-
-        // Ensure engineFileName is added to the tracked files by Lucene's TrackingDirectoryWrapper
-        state.directory.createOutput(engineFileName, state.context).close();
-        indexCreator.createIndex();
-        writeFooter(indexPath, engineFileName);
-    }
-
-    private void recordMergeStats(int length, long arraySize) {
-        KNNGraphValue.MERGE_CURRENT_OPERATIONS.decrement();
-        KNNGraphValue.MERGE_CURRENT_DOCS.decrementBy(length);
-        KNNGraphValue.MERGE_CURRENT_SIZE_IN_BYTES.decrementBy(arraySize);
-        KNNGraphValue.MERGE_TOTAL_OPERATIONS.increment();
-        KNNGraphValue.MERGE_TOTAL_DOCS.incrementBy(length);
-        KNNGraphValue.MERGE_TOTAL_SIZE_IN_BYTES.incrementBy(arraySize);
-    }
-
-    private void recordRefreshStats() {
-        KNNGraphValue.REFRESH_TOTAL_OPERATIONS.increment();
-    }
-
-    private void createKNNIndexFromTemplate(Model model, KNNCodecUtil.Pair pair, KNNEngine knnEngine, String indexPath) {
-        Map<String, Object> parameters = new HashMap<>();
-        parameters.put(KNNConstants.INDEX_THREAD_QTY, KNNSettings.state().getSettingValue(KNNSettings.KNN_ALGO_PARAM_INDEX_THREAD_QTY));
-
-        IndexUtil.updateVectorDataTypeToParameters(parameters, model.getModelMetadata().getVectorDataType());
-
-        AccessController.doPrivileged((PrivilegedAction<Void>) () -> {
-            JNIService.createIndexFromTemplate(
-                pair.docs,
-                pair.getVectorAddress(),
-                pair.getDimension(),
-                indexPath,
-                model.getModelBlob(),
-                parameters,
-                knnEngine
-            );
-            return null;
-        });
-    }
-
-    private void createKNNIndexFromScratch(FieldInfo fieldInfo, KNNCodecUtil.Pair pair, KNNEngine knnEngine, String indexPath)
-        throws IOException {
-        Map<String, Object> parameters = new HashMap<>();
-        Map<String, String> fieldAttributes = fieldInfo.attributes();
-        String parametersString = fieldAttributes.get(PARAMETERS);
-        // parametersString will be null when legacy mapper is used
-        if (parametersString == null) {
-            parameters.put(KNNConstants.SPACE_TYPE, fieldAttributes.getOrDefault(KNNConstants.SPACE_TYPE, SpaceType.DEFAULT.getValue()));
-
-            String efConstruction = fieldAttributes.get(KNNConstants.HNSW_ALGO_EF_CONSTRUCTION);
-            Map<String, Object> algoParams = new HashMap<>();
-            if (efConstruction != null) {
-                algoParams.put(KNNConstants.METHOD_PARAMETER_EF_CONSTRUCTION, Integer.parseInt(efConstruction));
-            }
-
-            String m = fieldAttributes.get(KNNConstants.HNSW_ALGO_M);
-            if (m != null) {
-                algoParams.put(KNNConstants.METHOD_PARAMETER_M, Integer.parseInt(m));
-            }
-            parameters.put(PARAMETERS, algoParams);
-        } else {
-            parameters.putAll(
-                XContentHelper.createParser(
-                    NamedXContentRegistry.EMPTY,
-                    DeprecationHandler.THROW_UNSUPPORTED_OPERATION,
-                    new BytesArray(parametersString),
-                    MediaTypeRegistry.getDefaultMediaType()
-                ).map()
-            );
-        }
-
-        // In OpenSearch 2.16, we added the prefix for binary indices in the index description in the codec logic.
-        // After 2.16, we added the binary prefix in the faiss library code. However, to ensure backwards compatibility,
-        // we need to ensure that if the description does not contain the prefix but the type is binary, we add the
-        // description.
-        maybeAddBinaryPrefixForFaissBWC(knnEngine, parameters, fieldAttributes);
-
-        // Used to determine how many threads to use when indexing
-        parameters.put(KNNConstants.INDEX_THREAD_QTY, KNNSettings.state().getSettingValue(KNNSettings.KNN_ALGO_PARAM_INDEX_THREAD_QTY));
-
-        // Pass the path for the nms library to save the file
-        AccessController.doPrivileged((PrivilegedAction<Void>) () -> {
-            JNIService.createIndex(pair.docs, pair.getVectorAddress(), pair.getDimension(), indexPath, parameters, knnEngine);
-            return null;
-        });
-    }
-
-    private void maybeAddBinaryPrefixForFaissBWC(KNNEngine knnEngine, Map<String, Object> parameters, Map<String, String> fieldAttributes) {
-        if (KNNEngine.FAISS != knnEngine) {
-            return;
-        }
-
-        if (!VectorDataType.BINARY.getValue()
-            .equals(fieldAttributes.getOrDefault(KNNConstants.VECTOR_DATA_TYPE_FIELD, VectorDataType.DEFAULT.getValue()))) {
-            return;
-        }
-
-        if (parameters.get(KNNConstants.INDEX_DESCRIPTION_PARAMETER) == null) {
-            return;
-        }
-
-        if (parameters.get(KNNConstants.INDEX_DESCRIPTION_PARAMETER).toString().startsWith(FAISS_BINARY_INDEX_DESCRIPTION_PREFIX)) {
-            return;
-        }
-
-        parameters.put(
-            KNNConstants.INDEX_DESCRIPTION_PARAMETER,
-            FAISS_BINARY_INDEX_DESCRIPTION_PREFIX + parameters.get(KNNConstants.INDEX_DESCRIPTION_PARAMETER).toString()
-        );
-        IndexUtil.updateVectorDataTypeToParameters(parameters, VectorDataType.BINARY);
+        NativeIndexWriter.getWriter(field).createKNNIndex(field, valuesProducer, state, isMerge, isRefresh);
     }
 
     /**
@@ -335,53 +130,5 @@ class KNN80DocValuesConsumer extends DocValuesConsumer implements Closeable {
     @Override
     public void close() throws IOException {
         delegatee.close();
-    }
-
-    @FunctionalInterface
-    private interface NativeIndexCreator {
-        void createIndex() throws IOException;
-    }
-
-    private void writeFooter(String indexPath, String engineFileName) throws IOException {
-        // Opens the engine file that was created and appends a footer to it. The footer consists of
-        // 1. A Footer magic number (int - 4 bytes)
-        // 2. A checksum algorithm id (int - 4 bytes)
-        // 3. A checksum (long - bytes)
-        // The checksum is computed on all the bytes written to the file up to that point.
-        // Logic where footer is written in Lucene can be found here:
-        // https://github.com/apache/lucene/blob/branch_9_0/lucene/core/src/java/org/apache/lucene/codecs/CodecUtil.java#L390-L412
-        OutputStream os = Files.newOutputStream(Paths.get(indexPath), StandardOpenOption.APPEND);
-        ByteBuffer byteBuffer = ByteBuffer.allocate(8).order(ByteOrder.BIG_ENDIAN);
-        byteBuffer.putInt(FOOTER_MAGIC);
-        byteBuffer.putInt(0);
-        os.write(byteBuffer.array());
-        os.flush();
-
-        ChecksumIndexInput checksumIndexInput = state.directory.openChecksumInput(engineFileName, state.context);
-        checksumIndexInput.seek(checksumIndexInput.length());
-        long value = checksumIndexInput.getChecksum();
-        checksumIndexInput.close();
-
-        if (isChecksumValid(value)) {
-            throw new IllegalStateException("Illegal CRC-32 checksum: " + value + " (resource=" + os + ")");
-        }
-
-        // Write the CRC checksum to the end of the OutputStream and close the stream
-        byteBuffer.putLong(0, value);
-        os.write(byteBuffer.array());
-        os.close();
-    }
-
-    private boolean isChecksumValid(long value) {
-        // Check pulled from
-        // https://github.com/apache/lucene/blob/branch_9_0/lucene/core/src/java/org/apache/lucene/codecs/CodecUtil.java#L644-L647
-        return (value & CRC32_CHECKSUM_SANITY) != 0;
-    }
-
-    private VectorTransfer getVectorTransfer(VectorDataType vectorDataType) {
-        if (VectorDataType.BINARY == vectorDataType) {
-            return new VectorTransferByte(KNNSettings.getVectorStreamingMemoryLimit().getBytes());
-        }
-        return new VectorTransferFloat(KNNSettings.getVectorStreamingMemoryLimit().getBytes());
     }
 }