+++ src/test/java/org/opensearch/knn/jni/JNIServiceTests.java
@@ -86,7 +86,7 @@ public class JNIServiceTests extends KNNTestCase {
     public void testCreateIndex_invalid_engineNotSupported() {
         expectThrows(
             IllegalArgumentException.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 new int[] {},
                 0,
                 0,
@@ -100,21 +100,14 @@ public class JNIServiceTests extends KNNTestCase {
     public void testCreateIndex_invalid_engineNull() {
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
-                new int[] {},
-                0,
-                0,
-                "test",
-                ImmutableMap.of(KNNConstants.SPACE_TYPE, SpaceType.L2.getValue()),
-                null
-            )
+            () -> TestUtils.createIndex(new int[] {}, 0, 0, "test", ImmutableMap.of(KNNConstants.SPACE_TYPE, SpaceType.L2.getValue()), null)
         );
     }
 
     public void testCreateIndex_nmslib_invalid_noSpaceType() {
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 testData.indexData.docs,
                 testData.loadDataToMemoryAddress(),
                 testData.indexData.getDimension(),
@@ -133,7 +126,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile1 = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress,
                 vectors1[0].length,
@@ -149,7 +142,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile2 = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress2,
                 vectors2[0].length,
@@ -168,7 +161,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 null,
                 memoryAddress,
                 0,
@@ -180,7 +173,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 0,
                 0,
@@ -192,7 +185,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress,
                 0,
@@ -204,12 +197,12 @@ public class JNIServiceTests extends KNNTestCase {
 
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(docIds, memoryAddress, 0, tmpFile.toAbsolutePath().toString(), null, KNNEngine.NMSLIB)
+            () -> TestUtils.createIndex(docIds, memoryAddress, 0, tmpFile.toAbsolutePath().toString(), null, KNNEngine.NMSLIB)
         );
 
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress,
                 0,
@@ -228,7 +221,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress,
                 vectors[0].length,
@@ -254,7 +247,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress,
                 vectors[0].length,
@@ -274,7 +267,7 @@ public class JNIServiceTests extends KNNTestCase {
 
             Path tmpFile = createTempFile();
 
-            JNIService.createIndex(
+            TestUtils.createIndex(
                 testData.indexData.docs,
                 testData.loadDataToMemoryAddress(),
                 testData.indexData.getDimension(),
@@ -286,7 +279,7 @@ public class JNIServiceTests extends KNNTestCase {
 
             tmpFile = createTempFile();
 
-            JNIService.createIndex(
+            TestUtils.createIndex(
                 testData.indexData.docs,
                 testData.loadDataToMemoryAddress(),
                 testData.indexData.getDimension(),
@@ -310,7 +303,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 testData.loadDataToMemoryAddress(),
                 testData.indexData.getDimension(),
@@ -329,7 +322,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile1 = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress,
                 vectors1[0].length,
@@ -344,7 +337,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile2 = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress,
                 vectors2[0].length,
@@ -364,7 +357,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 null,
                 memoryAddress,
                 0,
@@ -376,7 +369,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 0,
                 0,
@@ -388,7 +381,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 testData.loadDataToMemoryAddress(),
                 testData.indexData.getDimension(),
@@ -400,7 +393,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 testData.loadDataToMemoryAddress(),
                 testData.indexData.getDimension(),
@@ -412,7 +405,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 testData.loadDataToMemoryAddress(),
                 testData.indexData.getDimension(),
@@ -432,7 +425,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress,
                 vectors[0].length,
@@ -452,7 +445,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress,
                 vectors[0].length,
@@ -470,7 +463,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress,
                 vectors[0].length,
@@ -493,7 +486,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 memoryAddress,
                 vectors[0].length,
@@ -517,7 +510,7 @@ public class JNIServiceTests extends KNNTestCase {
         String sqfp16IndexDescription = "HNSW16,SQfp16";
         long memoryAddress = JNICommons.storeVectorData(0, vectors, (long) vectors.length * vectors[0].length);
         Path tmpFile = createTempFile();
-        JNIService.createIndex(
+        TestUtils.createIndex(
             docIds,
             memoryAddress,
             vectors[0].length,
@@ -540,7 +533,7 @@ public class JNIServiceTests extends KNNTestCase {
         float[][] truncatedVectors = truncateToFp16Range(testData.indexData.vectors);
         long memoryAddress = JNICommons.storeVectorData(0, truncatedVectors, (long) truncatedVectors.length * truncatedVectors[0].length);
         Path tmpFile = createTempFile();
-        JNIService.createIndex(
+        TestUtils.createIndex(
             testData.indexData.docs,
             memoryAddress,
             testData.indexData.getDimension(),
@@ -634,7 +627,7 @@ public class JNIServiceTests extends KNNTestCase {
         Path tmpFile = createTempFile();
         expectThrows(
             Exception.class,
-            () -> JNIService.createIndex(
+            () -> TestUtils.createIndex(
                 docIds,
                 testData.loadDataToMemoryAddress(),
                 testData.indexData.getDimension(),
@@ -660,7 +653,7 @@ public class JNIServiceTests extends KNNTestCase {
         for (String method : methods) {
             for (SpaceType spaceType : spaces) {
                 Path tmpFile1 = createTempFile();
-                JNIService.createIndex(
+                TestUtils.createIndex(
                     testData.indexData.docs,
                     testData.loadDataToMemoryAddress(),
                     testData.indexData.getDimension(),
@@ -677,7 +670,7 @@ public class JNIServiceTests extends KNNTestCase {
     public void testCreateIndex_binary_faiss_valid() {
         Path tmpFile1 = createTempFile();
         long memoryAddr = testData.loadBinaryDataToMemoryAddress();
-        JNIService.createIndex(
+        TestUtils.createIndex(
             testData.indexData.docs,
             memoryAddr,
             testData.indexData.getDimension(),
@@ -733,7 +726,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         Path tmpFile = createTempFile();
 
-        JNIService.createIndex(
+        TestUtils.createIndex(
             testData.indexData.docs,
             testData.loadDataToMemoryAddress(),
             testData.indexData.getDimension(),
@@ -769,7 +762,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         Path tmpFile = createTempFile();
 
-        JNIService.createIndex(
+        TestUtils.createIndex(
             testData.indexData.docs,
             testData.loadDataToMemoryAddress(),
             testData.indexData.getDimension(),
@@ -799,7 +792,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         Path tmpFile = createTempFile();
 
-        JNIService.createIndex(
+        TestUtils.createIndex(
             testData.indexData.docs,
             testData.loadDataToMemoryAddress(),
             testData.indexData.getDimension(),
@@ -829,7 +822,7 @@ public class JNIServiceTests extends KNNTestCase {
 
             Path tmpFile = createTempFile();
 
-            JNIService.createIndex(
+            TestUtils.createIndex(
                 testData.indexData.docs,
                 testData.loadDataToMemoryAddress(),
                 testData.indexData.getDimension(),
@@ -862,7 +855,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         Path tmpFile = createTempFile();
 
-        JNIService.createIndex(
+        TestUtils.createIndex(
             testData.indexData.docs,
             testData.loadDataToMemoryAddress(),
             testData.indexData.getDimension(),
@@ -888,7 +881,7 @@ public class JNIServiceTests extends KNNTestCase {
         for (String method : methods) {
             for (SpaceType spaceType : spaces) {
                 Path tmpFile = createTempFile();
-                JNIService.createIndex(
+                TestUtils.createIndex(
                     testData.indexData.docs,
                     testData.loadDataToMemoryAddress(),
                     testData.indexData.getDimension(),
@@ -949,7 +942,7 @@ public class JNIServiceTests extends KNNTestCase {
         for (String method : methods) {
             for (SpaceType spaceType : spaces) {
                 Path tmpFile = createTempFile();
-                JNIService.createIndex(
+                TestUtils.createIndex(
                     testDataNested.indexData.docs,
                     testData.loadDataToMemoryAddress(),
                     testDataNested.indexData.getDimension(),
@@ -992,7 +985,7 @@ public class JNIServiceTests extends KNNTestCase {
         for (String method : methods) {
             Path tmpFile = createTempFile();
             long memoryAddr = testData.loadBinaryDataToMemoryAddress();
-            JNIService.createIndex(
+            TestUtils.createIndex(
                 testData.indexData.docs,
                 memoryAddr,
                 testData.indexData.getDimension(),
@@ -1071,7 +1064,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         Path tmpFile = createTempFile();
 
-        JNIService.createIndex(
+        TestUtils.createIndex(
             testData.indexData.docs,
             testData.loadDataToMemoryAddress(),
             testData.indexData.getDimension(),
@@ -1095,7 +1088,7 @@ public class JNIServiceTests extends KNNTestCase {
 
         Path tmpFile = createTempFile();
 
-        JNIService.createIndex(
+        TestUtils.createIndex(
             testData.indexData.docs,
             testData.loadDataToMemoryAddress(),
             testData.indexData.getDimension(),
@@ -1235,7 +1228,7 @@ public class JNIServiceTests extends KNNTestCase {
         return trainPointer1;
     }
 
-    public void testCreateIndexFromTemplate() throws IOException {
+    public void createIndexFromTemplate() throws IOException {
 
         long trainPointer1 = JNIService.transferVectors(0, testData.indexData.vectors);
         assertNotEquals(0, trainPointer1);
@@ -1445,7 +1438,7 @@ public class JNIServiceTests extends KNNTestCase {
 
     private String createFaissHNSWIndex(SpaceType spaceType) throws IOException {
         Path tmpFile = createTempFile();
-        JNIService.createIndex(
+        TestUtils.createIndex(
             testData.indexData.docs,
             testData.loadDataToMemoryAddress(),
             testData.indexData.getDimension(),